starting to modify p-id to f-id
There is a relation between the f-id and p-id. A requested requirement for p-id's is that we would like the p-id to increment with every version of an article. The p-id is labelled in the article XML, and as this id will actually increment for every version of the xml, and the file number f-id won't we have renamed this part of the naming convention to remove a conflict in the name.